+++ OptSQPSR1.cpp
@@ -346,7 +346,7 @@ void OptSQPSR1::computeGradient(const double *x, double f0, double *aGrad)
 #endif // SCALE_OPT_VARIABLES_SR1
 	
 	// branch lengths
-	for(i=0; i<mNumTimes; ++i)
+	for(i=0; i<static_cast<size_t>(mNumTimes); ++i)
 	{
 		eh = sqrt_eps * ( 1.0 + x_[i] );
 		if( x_[i] + eh > mUpperBoundUnscaled[i] )
@@ -355,7 +355,7 @@ void OptSQPSR1::computeGradient(const double *x, double f0, double *aGrad)
 		delta[i] = mXEvaluator[i] - x_[i];
 	}
 
-	for(i=0; i<mNumTimes; ++i)
+	for(i=0; i<static_cast<size_t>(mNumTimes); ++i)
 	{
 		if(mActiveSet[i] == 0)
 		{
@@ -367,7 +367,7 @@ void OptSQPSR1::computeGradient(const double *x, double f0, double *aGrad)
 	
 	// other variables
 	memcpy(&mXEvaluator[0], x_, size_vect);
-	for(; i<mN; ++i)
+	for(; i<static_cast<size_t>(mN); ++i)
 	{
 		if (mActiveSet[i] == 0)
 		{
@@ -479,7 +479,7 @@ void OptSQPSR1::lineSearch(double *aalpha, double *x, double *f)
 	
 	// begin by a backtrace
 	size_t iter = 0;
-	while(phi > phi_0 + phi_0_prime*a*c1 && iter < maxIterBack)
+	while(phi > phi_0 + phi_0_prime*a*c1 && iter < static_cast<size_t>(maxIterBack))
 	{
 		++iter;
 		a_prev = a;
@@ -500,7 +500,7 @@ void OptSQPSR1::lineSearch(double *aalpha, double *x, double *f)
 	if (phi_a_prime < 0.0 && fabs(a - *aalpha) > 1e-8)
 	{
 		double a0 = a_prev;
-		while(phi < phi_prev && iter < maxIterBack)
+		while(phi < phi_prev && iter < static_cast<size_t>(maxIterBack))
 		{
 			++iter;
 			
@@ -520,7 +520,7 @@ void OptSQPSR1::lineSearch(double *aalpha, double *x, double *f)
 	else
 	{
 		sigma_bas = 0.7;
-		while(phi < phi_prev && iter < maxIterUp)
+		while(phi < phi_prev && iter < static_cast<size_t>(maxIterUp))
 		{
 			++iter;
